Try much harder to guess a sensible context and distribution
In particular, it tries to use the current context to guess the
distribution. Because they depend on each other, it is tricky not to get things
messed up. Therefore, I tried hard adding useful debugging tips and preventively
raise an error when one is finally not guessed correctly.

Also, two commands are added to easily ask what values are eventually used.