fix: Make sure we use the appropriate IP to connect to the registry
In a default local env, the bridge network is 172.17.0.1, but in
nested docker envs, it changes.

For instance, in the CI, with k8s in docker in buildkit-runc in
docker, it becomes 172.20.0.1.

In that case, binding on 172.17.0.1 may work, but it will conflict
with a stack already running on the same computer.

Now, we compute the appropriate ip by inspecting the docker bridge
network when needed.

Also, the earthly configuration to deal with the registry should
not be updated at install time, but rather when the local registry
is started, because only now we can compute the ip that will
actually be used to discuss with it.