refactor: Wrap the helm errors to avoid frightening the user
Helm errors are hard to read and counter intuitive. Thus, wrapping
them allow providing more context and indicating what to do next.

For instance, most of the errors we face today are time out due to
temporary overload on the machine during the installation. After
getting the context deadline exceeded, we can indicate that the
user can try running it again before freaking out.

In order not to break the current behavior of silent_call, a version of it that
raises a SilentCallFailed error was separated so that we can make us of it is
helm_install, while silent_call itself kept the behavior of exiting directly.