fix: Make even more sure that the earthly configuration will be correctly adapted
We used simple substring matching to decide whether we would add
the configuration or not.

In case there is already a configuration in place, it would tend
to generate a bad toml content.

Now, we parse and dump toml using third party libraries so that we
make more sure that the content is appropriately parsed and
appropriately generated afterwards.